Restored compilation unit test and moved in other test class
