+++ rpgJavaInterpreter-core/src/test/resources/HELLO.json
@@ -1,1 +1,1 @@
-{"fileDefinitions":[],"dataDefinitions":[{"key":2,"name":"Msg","dataDefType":{"type":"com.smeup.rpgparser.interpreter.StringType","length":12},"position":{"start":{"line":1,"column":5},"end":{"line":1,"column":81}}}],"main":{"stmts":[{"type":"com.smeup.rpgparser.parsing.ast.EvalStmt","target":{"type":"com.smeup.rpgparser.parsing.ast.DataRefExpr","variable":{"name":"Msg"},"position":{"start":{"line":2,"column":35},"end":{"line":2,"column":38}}},"expression":{"type":"com.smeup.rpgparser.parsing.ast.StringLiteral","value":"Hello World!","position":{"start":{"line":2,"column":42},"end":{"line":2,"column":56}}},"position":{"start":{"line":2,"column":25},"end":{"line":2,"column":81}}},{"type":"com.smeup.rpgparser.parsing.ast.DisplayStmt","factor1":null,"response":{"type":"com.smeup.rpgparser.parsing.ast.DataRefExpr","variable":{"name":"Msg"},"position":{"start":{"line":3,"column":49},"end":{"line":3,"column":52}}},"position":{"start":{"line":3,"column":25},"end":{"line":3,"column":81}}},{"type":"com.smeup.rpgparser.parsing.ast.SetStmt","valueSet":"ON","indicators":[{"type":"com.smeup.rpgparser.parsing.ast.DataWrapUpIndicatorExpr","dataWrapUpChoice":"LR","position":{"start":{"line":4,"column":35},"end":{"line":4,"column":81}}}],"position":{"start":{"line":4,"column":25},"end":{"line":4,"column":81}}}],"position":{"start":{"line":2,"column":25},"end":{"line":4,"column":81}}},"subroutines":[],"compileTimeArrays":[],"directives":[],"position":{"start":{"line":1,"column":5},"end":{"line":4,"column":81}}}
+{"fileDefinitions":[],"dataDefinitions":[{"key":0,"name":"Msg","dataDefType":{"type":"com.smeup.rpgparser.interpreter.StringType","length":12},"position":{"start":{"line":1,"column":5},"end":{"line":1,"column":82}}}],"main":{"stmts":[{"type":"com.smeup.rpgparser.parsing.ast.EvalStmt","target":{"type":"com.smeup.rpgparser.parsing.ast.DataRefExpr","variable":{"name":"Msg"},"position":{"start":{"line":2,"column":35},"end":{"line":2,"column":38}}},"expression":{"type":"com.smeup.rpgparser.parsing.ast.StringLiteral","value":"Hello World!","position":{"start":{"line":2,"column":42},"end":{"line":2,"column":56}}},"position":{"start":{"line":2,"column":25},"end":{"line":2,"column":82}}},{"type":"com.smeup.rpgparser.parsing.ast.DisplayStmt","factor1":null,"response":{"type":"com.smeup.rpgparser.parsing.ast.DataRefExpr","variable":{"name":"Msg"},"position":{"start":{"line":3,"column":49},"end":{"line":3,"column":52}}},"position":{"start":{"line":3,"column":25},"end":{"line":3,"column":82}}},{"type":"com.smeup.rpgparser.parsing.ast.SetStmt","valueSet":"ON","indicators":[{"type":"com.smeup.rpgparser.parsing.ast.DataWrapUpIndicatorExpr","dataWrapUpChoice":"LR","position":{"start":{"line":4,"column":35},"end":{"line":4,"column":82}}}],"position":{"start":{"line":4,"column":25},"end":{"line":4,"column":82}}}],"position":{"start":{"line":2,"column":25},"end":{"line":4,"column":82}}},"subroutines":[],"compileTimeArrays":[],"directives":[],"position":{"start":{"line":1,"column":5},"end":{"line":4,"column":82}}}
